requirements.txt
aiogram==3.11.0
python-decouple~=3.8
telebot~=0.0.5
requests~=2.31.0
langchain~=0.3.9
transformers~=4.33.2
torch~=2.0.1
sentence-transformers~=2.2.2
nltk~=3.8.1
faiss-cpu~=1.7.4
faiss-gpu~=1.7.4
python-docx~=0.8.11
python-pptx~=0.6.21
pytesseract~=0.3.13
pillow~=9.5.0
pymupdf~=1.0.0
camelot-py[cv]~=0.10.1
huggingface_hub~=0.13.3
qwen-vl-utils~=0.1.0  # Убедитесь, что версия актуальна
flash-attn~=1.0.0
optimum~=1.0.0
auto-gptq~=0.0.1
bitsandbytes~=0.40.0
byaldi~=0.1.0
claudette~=0.1.0
pdf2image~=1.16.0
